requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
ftfy==6.3.1
lxml==6.0.2
numpy
# numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
pillow
# pillow==12.0.0
pycparser==2.23
PyMuPDF==1.26.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
Sastrawi==1.0.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
wcwidth==0.2.14